+++ test/e2e/E2E.test.ts
@@ -1,11 +1,11 @@
-import {withTempDirectory} from '../../test/helpers/directories';
-import * as util from '../../util';
-import {mocked} from 'ts-jest/utils';
+import { mocked } from 'ts-jest/utils';
 import * as path from 'path';
-import {fs} from '../../p/fs';
-import { E2E } from "../../commands/e2e";
+import * as fs from "fs";
+import { E2E } from "../../src/commands/e2e";
+import { withTempDirectory } from "../helpers/directories";
+import * as util from "../../src/util";
 
-jest.mock('../../util');
+jest.mock('../../src/util');
 
 test('E2E detects Allure Reporter in Dev Dependencies', async () => {
     await withTempDirectory('e2e-allure', async (dir) => {
@@ -20,7 +20,7 @@ test('E2E detects Allure Reporter in Dev Dependencies', async () => {
         };
 
         const packageJsonPath = path.join(dir, 'package.json');
-        await fs.writeFileAsync(packageJsonPath, JSON.stringify(packageJson), 'utf8');
+        fs.writeFileSync(packageJsonPath, JSON.stringify(packageJson), 'utf8');
 
         const e2eCommand = new E2E();
         e2eCommand.prepareAndMayExecute({});
@@ -44,7 +44,7 @@ test('E2E detects Allure Reporter in Dependencies', async () => {
         };
 
         const packageJsonPath = path.join(dir, 'package.json');
-        await fs.writeFileAsync(packageJsonPath, JSON.stringify(packageJson), 'utf8');
+        fs.writeFileSync(packageJsonPath, JSON.stringify(packageJson), 'utf8');
 
         const e2eCommand = new E2E();
         e2eCommand.prepareAndMayExecute({});
@@ -65,7 +65,7 @@ test('E2E detects missing Allure Reporter', async () => {
         };
 
         const packageJsonPath = path.join(dir, 'package.json');
-        await fs.writeFileAsync(packageJsonPath, JSON.stringify(packageJson), 'utf8');
+        fs.writeFileSync(packageJsonPath, JSON.stringify(packageJson), 'utf8');
 
         const e2eCommand = new E2E();
         e2eCommand.prepareAndMayExecute({});
@@ -86,7 +86,7 @@ test('E2E detects WDIO Image-Comparison-Service in Dev Dependencies', async () =
         };
 
         const packageJsonPath = path.join(dir, 'package.json');
-        await fs.writeFileAsync(packageJsonPath, JSON.stringify(packageJson), 'utf8');
+        fs.writeFileSync(packageJsonPath, JSON.stringify(packageJson), 'utf8');
 
         const e2eCommand = new E2E();
         e2eCommand.prepareAndMayExecute({});
@@ -110,7 +110,7 @@ test('E2E detects WDIO Image-Comparison-Service in Dependencies', async () => {
         };
 
         const packageJsonPath = path.join(dir, 'package.json');
-        await fs.writeFileAsync(packageJsonPath, JSON.stringify(packageJson), 'utf8');
+        fs.writeFileSync(packageJsonPath, JSON.stringify(packageJson), 'utf8');
 
         const e2eCommand = new E2E();
         e2eCommand.prepareAndMayExecute({});
@@ -131,7 +131,7 @@ test('E2E detects missing WDIO Image-Comparison-Service', async () => {
         };
 
         const packageJsonPath = path.join(dir, 'package.json');
-        await fs.writeFileAsync(packageJsonPath, JSON.stringify(packageJson), 'utf8');
+        fs.writeFileSync(packageJsonPath, JSON.stringify(packageJson), 'utf8');
 
         const e2eCommand = new E2E();
         e2eCommand.prepareAndMayExecute({});
