+++ test/helpers/gradle.ts
@@ -1,7 +1,6 @@
-import {withTempDirectory} from './directories';
-import * as path from 'path';
-import {fs} from '../../p/fs';
-import {promiseAllSettledParallel} from '../../promiseAllSettled';
+import { withTempDirectory } from "./directories";
+import * as path from "path";
+import * as fs from 'fs';
 
 export function withTempGradleBuild(func: (directory: string) => Promise<void>,
                                     buildGradleContent?: () => string,
@@ -20,10 +19,6 @@ export async function createGradleBuild(gradleDirectory: string,
                                         settingsGradleContent?: () => string): Promise<void> {
     const buildGradle = path.join(gradleDirectory, 'build.gradle');
     const settingsGradle = path.join(gradleDirectory, 'settings.gradle');
-    await promiseAllSettledParallel(
-        [
-            fs.writeFileAsync(buildGradle, buildGradleContent ? buildGradleContent() : '', 'utf-8'),
-            fs.writeFileAsync(settingsGradle, settingsGradleContent ? settingsGradleContent() : '', 'utf-8')
-        ]
-    );
+            fs.writeFileSync(buildGradle, buildGradleContent ? buildGradleContent() : '', 'utf-8'),
+            fs.writeFileSync(settingsGradle, settingsGradleContent ? settingsGradleContent() : '', 'utf-8')
 }
