+++ test/helpers/remoteRepositories.ts
@@ -1,9 +1,9 @@
-import {IReposDescriptor} from '../../commands/repos/models';
 import * as fs from 'fs';
 import * as path from 'path';
 import {withTempDirectory} from './directories';
 import * as child_process from 'child_process';
-import {enforceNewline} from '../../util';
+import { enforceNewline } from "../../src/util";
+import { IReposDescriptor } from "../../src/commands/repos/models";
 
 export function testWith(testSetupData: ITestSetupData): ITestRun {
     return new EvaluateWithRemoteRepos(testSetupData);
